refactor(keyboards): разбил файл keyboards.py на несколько модулей
